refactor(components): move auth components to src/components/auth
