+++ frontend/src/components/auth/auth-page/AuthPage.scss
