+++ frontend/src/pages/Login.tsx
@@ -25,7 +25,7 @@ import { useLocation, useNavigate } from "react-router-dom";
 import Alert from "react-bootstrap/Alert";
 import Form from "react-bootstrap/Form";
 
-import AuthPage from "@/components/AuthPage";
+import AuthPage from "@/components/auth/auth-page/AuthPage";
 import Button from "@/components/ui/button/Button";
 import Spinner from "@/components/ui/spinner/Spinner";
 import Stack from "@/components/ui/stack/Stack";
